ref(core): Remove `supportsDOMError`, `supportsHistory`, `supportsReportingObserver` exports
