+++ packages/core/src/utils/supports.ts
@@ -22,24 +22,6 @@ export function supportsErrorEvent(): boolean {
 }
 
 /**
- * Tells whether current environment supports DOMError objects
- * {@link supportsDOMError}.
- *
- * @returns Answer to the given question.
- */
-export function supportsDOMError(): boolean {
-  try {
-    // Chrome: VM89:1 Uncaught TypeError: Failed to construct 'DOMError':
-    // 1 argument required, but only 0 present.
-    // @ts-expect-error It really needs 1 argument, not 0.
-    new DOMError('');
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-/**
  * Tells whether current environment supports DOMException objects
  * {@link supportsDOMException}.
  *
@@ -52,16 +34,6 @@ export function supportsDOMException(): boolean {
   } catch {
     return false;
   }
-}
-
-/**
- * Tells whether current environment supports History API
- * {@link supportsHistory}.
- *
- * @returns Answer to the given question.
- */
-export function supportsHistory(): boolean {
-  return 'history' in WINDOW && !!WINDOW.history;
 }
 
 /**
@@ -139,16 +111,6 @@ export function supportsNativeFetch(): boolean {
   }
 
   return result;
-}
-
-/**
- * Tells whether current environment supports ReportingObserver API
- * {@link supportsReportingObserver}.
- *
- * @returns Answer to the given question.
- */
-export function supportsReportingObserver(): boolean {
-  return 'ReportingObserver' in WINDOW;
 }
 
 /**
